process_files_cloud_function/process_files_cloud_function/requirements-test.txt
pytest==7.2.0
google-cloud-contentwarehouse == 0.4.0
google-cloud-documentai == 2.20.0
google-cloud-storage == 2.7.0
google-cloud-logging == 3.5.0